Rename from visibility to connection
